requirements.txt
alembic==1.1.0
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
importlib-metadata==0.22
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.0.3
marshmallow-sqlalchemy==0.19.0
more-itertools==7.2.0
packaging==19.1
pluggy==0.13.0
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.2
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
wcwidth==0.1.7
Werkzeug==0.15.6
zipp==0.6.0
